refacto code, change url endpoint, update auth and notes handler

+++ controllers/api/deleteNoteHandler.js
@@ -1,19 +1,22 @@
 import response from '../../helpers/response.js';
-import { Notes } from '../../model/model.js';
+import { Notes, User } from '../../model/model.js';
 
 const deleteNoteHandler = async (req, res) => {
-  const { name } = req.user;
-  const { id } = req.params;
-
   try {
-    const foundNotes = await Notes.find({ username: name });
+    const { userId } = req.user;
+    const { id } = req.params;
+
+    const user = await User.findById(userId);
+    if (user.id !== userId) return response(403, 'fail', { message: 'Forbidden for user' }, res);
+
+    const foundNotes = await Notes.findOne({ userId });
 
     if (!foundNotes) throw Error();
 
     const result = await Notes.deleteOne({ _id: id });
 
     if (result.deletedCount) {
-      return response(200, 'success', { message: 'Notes berhasil dihapus' }, res);
+      return response(200, 'success', { message: 'Notes berhasil dihapus', data: { noteId: foundNotes.id } }, res);
     }
 
     throw Error();
